Restructure project files: separate app, routes, and models
